+++ src/BioProfiling.jl
@@ -34,6 +34,7 @@ export  logtransform,
         shuffled_distance_mahalanobis_median,
         distance_mahalanobis_median,
         robust_morphological_perturbation_value,
+        most_variable_features,
         umap
 using Statistics, StatsBase, DataFrames, Images, ImageMagick, UMAP, RCall, MultipleTesting
 using Distributed, ParallelDataTransfer
@@ -46,5 +47,6 @@ include("distances.jl")
 include("diagnostic.jl")
 include("visu.jl")
 include("rmpv.jl")
+include("interpret.jl")
 
 end # module