+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## v1.1.0 - unreleased
 ### Added
 - Option to remove R seed from distance computation
+- `most_variable_features` method
 
 ## v1.0.x - unreleased
 ### Changed